+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/baremetalmachine/_list.py
@@ -26,10 +26,10 @@ class List(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2026-07-01",
+        "version": "2026-08-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/providers/microsoft.networkcloud/baremetalmachines", "2026-07-01"],
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/baremetalmachines", "2026-07-01"],
+            ["mgmt-plane", "/subscriptions/{}/providers/microsoft.networkcloud/baremetalmachines", "2026-08-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/baremetalmachines", "2026-08-01-preview"],
         ]
     }
 
@@ -130,7 +130,7 @@ class List(AAZCommand):
                     "$top", self.ctx.args.top,
                 ),
                 **self.serialize_query_param(
-                    "api-version", "2026-07-01",
+                    "api-version", "2026-08-01-preview",
                     required=True,
                 ),
             }
@@ -592,6 +592,10 @@ class List(AAZCommand):
             )
 
             secret_archive_reference = cls._schema_on_200.value.Element.properties.secret_rotation_status.Element.secret_archive_reference
+            secret_archive_reference.encryption_public_key = AAZStrType(
+                serialized_name="encryptionPublicKey",
+                flags={"read_only": True},
+            )
             secret_archive_reference.key_vault_id = AAZStrType(
                 serialized_name="keyVaultId",
                 flags={"read_only": True},
@@ -687,7 +691,7 @@ class List(AAZCommand):
                     "$top", self.ctx.args.top,
                 ),
                 **self.serialize_query_param(
-                    "api-version", "2026-07-01",
+                    "api-version", "2026-08-01-preview",
                     required=True,
                 ),
             }
@@ -1149,6 +1153,10 @@ class List(AAZCommand):
             )
 
             secret_archive_reference = cls._schema_on_200.value.Element.properties.secret_rotation_status.Element.secret_archive_reference
+            secret_archive_reference.encryption_public_key = AAZStrType(
+                serialized_name="encryptionPublicKey",
+                flags={"read_only": True},
+            )
             secret_archive_reference.key_vault_id = AAZStrType(
                 serialized_name="keyVaultId",
                 flags={"read_only": True},