+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/virtualmachine/_create.py
@@ -23,9 +23,9 @@ class Create(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2026-07-01",
+        "version": "2026-08-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/virtualmachines/{}", "2026-07-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/virtualmachines/{}", "2026-08-01-preview"],
         ]
     }
 
@@ -356,9 +356,6 @@ class Create(AAZCommand):
             fmt=AAZStrArgFormat(
                 min_length=1,
             ),
-            blank=AAZPromptPasswordInput(
-                msg="Password:",
-            ),
         )
         vm_image_repository_credentials.registry_url = AAZStrArg(
             options=["registry-url"],
@@ -495,7 +492,7 @@ class Create(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2026-07-01",
+                    "api-version", "2026-08-01-preview",
                     required=True,
                 ),
             }