+++ src/networkcloud/azext_networkcloud/operations/common_commandoutputsettings.py
@@ -26,13 +26,10 @@ class CommandOutputSettings:
 
     @classmethod
     def pre_operations_update(cls, args):
-        if (
-            args.command_output_settings is None
-            or args.command_output_settings._data is None
-        ):
-            return
-
         if has_value(args.command_output_settings):
+            # skip validation when the value is explicitly set to null
+            if args.command_output_settings._data is None:
+                return
             # ensure container URL is provided
             if not has_value(args.command_output_settings.container_url):
                 raise RequiredArgumentMissingError(