+++ src/networkcloud/azext_networkcloud/tests/latest/test_l3network.py
@@ -112,7 +112,9 @@ def step_delete(test, checks=None):
     """L3Network delete operation"""
     if checks is None:
         checks = []
-    test.cmd("az networkcloud l3network delete --name {name} --resource-group {rg} -y")
+    test.cmd(
+        "az networkcloud l3network delete --name {name} --resource-group {rg} --yes"
+    )
 
 
 def step_list_resource_group(test, checks=None):